Added capability for additional text fields to be added to AlertViews
Added code to be able to handle multiple textfields in an alertview
simply by passing in placeholders for the desired textfields.
Placeholders should be blank (@"") if no placeholder is required.

An outlet has also been provided to be able to access the value in the
textfields from the block for the Done button.

There were a number of left over methods after these changes which I
could not see a use for, so I have removed them to avoid clutter /
unneccesary complexity